pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dpy-dialog"
version = "0.0.1"
authors = [
    { name = "Tanner B. Corcoran", email = "tannerbcorcoran@gmail.com" },
]
description = "A dialog handler for discord.py to make multipart dialog through Discord easy"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "discord.py >= 2.0.0"
]

[project.urls]
"Homepage" = "https://github.com/tanrbobanr/dpy-dialog"